Bump Go and golangci-lint
Pulling in the latest function-sdk-go bumps the Go toolchain in go.mod.
This bumps the Go version we use in CI to match.

It's unclear to me from a quick skim of the (lengthy) Go toolchain docs
if we could just omit our GO_VERSION variable now and let the toolchain
directive in go.mod sort it out.

Signed-off-by: Nic Cope <nicc@rk0n.org>